#733 split stats addons
